backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
scikit-image==0.22.0
shapely==2.0.2
geojson==3.1.0
rasterio==1.3.9
scipy==1.11.4
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
tqdm==4.66.1
matplotlib==3.8.2
